libap: FIONREAD's argument is an int*, and trace read() under APEXP_DEBUG
chatty9p settled where the vts shell's EOF comes from, by what is
absent: there is not one Tread on the shell's fid in the whole session,
and between the prompt and readline's ESC[?2004l there is no 9P message
at all. Reads plainly work -- ctl and ttyctl are read a moment later in
the same log. So bash did not ask vts for input and get a wrong answer;
it never asked. That takes vts out of it.

Three mechanisms read out of the source were refuted on the way:
FIONREAD-as-fstat (readline reaches it only through rl_gather_tyi, and
this bash never sets rl_event_hook), O_NONBLOCK diverting the read into
_readbuf (oflags is O_RDWR=2, O_NONBLOCK is 4, so isregular() is never
even called), and a stat from libap's own read path (same reason).

So instrument rather than argue further:

  - read() prints which branch it took and what it answered, including
    the two answers that never reach the kernel (EBADF and a zero
    count). Gated on a new _apdbgon(), so a normal run costs one load
    and one branch rather than a call per line.
  - vts passes APEXP_DEBUG to the shell whenever $vtsdebug is set.
    Those lines go to fd 2, which is the terminal, so they travel back
    through the server and land in vts's log beside the 9P trace.

And one real bug found on the way, recorded as found rather than
measured: ioctl(FIONREAD) wrote *(long*)arg where the argument is an
int* everywhere it is used, storing eight bytes on amd64 and smashing
the four beyond -- readline passes &chars_avail, an int local.

+++ sys/src/ape/lib/ap/misc/ioctl.c
@@ -22,13 +22,37 @@ ioctl(int fd, unsigned long request, void* arg)
 	int n;
 
 	switch(request) {
+	/*
+	 * FIONREAD'S ARGUMENT IS AN `int *', NOT A `long *'.
+	 *
+	 * BSD, Linux and every caller in this tree pass the address of an
+	 * `int'; readline's rl_gather_tyi() passes `&chars_avail', an int
+	 * local. Writing a `long' through it stores EIGHT bytes on amd64
+	 * and smashes the four beyond -- whatever the compiler put next in
+	 * the caller's frame. The store even succeeds, so nothing
+	 * complains; what shows up is the neighbouring variable being zero
+	 * for no reason, somewhere else entirely.
+	 *
+	 * Found while reading a vts trace, NOT measured: it is not what
+	 * made bash exit there (the EOF happened with no stat before it).
+	 * It is wrong on its own terms, which is enough.
+	 *
+	 * The ANSWER is still an approximation and worth saying out loud:
+	 * st_size is the number of bytes available only for a file that
+	 * has a size. For a terminal, a pipe or a socket Plan 9 reports 0,
+	 * so this says "nothing to read" always -- which is the safe
+	 * direction (a caller waits rather than reading what is not there)
+	 * but is not the truth. Plan 9's stat on a PIPE does report what
+	 * is queued, so a real answer is available for that case; it needs
+	 * its own round and a test, not a guess here.
+	 */
 	case FIONREAD:
 		if(fstat(fd, &d) < 0) {
 			errno = EBADF;
 			return -1;
 		}
 		/* this works if the file is buffered somehow */
-		*(long*)arg = d.st_size;
+		*(int*)arg = (int)d.st_size;
 		return 0;
 
 	case TIOCGWINSZ:
